Combine some of the more direct scanner generators
