+++ ulox/ulox.core/Package/Runtime/Scanner/Generators/IdentifierScannerTokenGenerator.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 using System.Text;
 
 namespace ULox
@@ -24,6 +25,7 @@ namespace ULox
                 (c >= 'A' && c <= 'Z') ||
                 c == '_';
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsAlphaNumber(int c) => IsAlpha(c) || NumberScannerTokenGenerator.IsDigit(c);
 
         public bool DoesMatchChar(char ch) => IsAlpha(ch);
