+++ ulox/ulox.core/Package/Runtime/Scanner/Generators/NumberScannerTokenGenerator.cs
@@ -1,4 +1,5 @@
-﻿using System.Text;
+﻿using System.Runtime.CompilerServices;
+using System.Text;
 
 namespace ULox
 {
@@ -6,6 +7,7 @@ namespace ULox
     {
         private readonly StringBuilder workingSpaceStringBuilder = new StringBuilder();
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static bool IsDigit(int ch) => ch >= '0' && ch <= '9';
 
         public bool DoesMatchChar(char ch) => IsDigit(ch);
