--- ulox/ulox.core/Package/Runtime/Scanner/Generators/WhiteSpaceScannerTokenGenerator.cs
@@ -1,21 +1,0 @@
-﻿namespace ULox
-{
-    public sealed class WhiteSpaceScannerTokenGenerator : IScannerTokenGenerator
-    {
-        public Token Consume(Scanner scanner) => Scanner.SharedNoToken;
-
-        public bool DoesMatchChar(char ch)
-        {
-            switch (ch)
-            {
-                case ' ':
-                case '\r':
-                case '\t':
-                case '\n':
-                return true;
-            }
-
-            return false;
-        }
-    }
-}
